Update GitHub Actions in code-validation.yaml
The commit updates the version of Node.js in the setup line from v2 to v4 in the GitHub Actions workflow file. It also updates the version of the actions/checkout and actions-gh-pages from v3 to v4. These changes ensure the code validation workflow uses the latest version of the necessary tools.